Can't test record deletion at expiry because it happens too slowly
